FLoC/requirements.txt
absl-py @ file:///C:/ci/absl-py_1623867344508/work
astor==0.8.1
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2021.5.30
click @ file:///tmp/build/80754af9/click_1621604852318/work
codetiming==1.3.0
coverage @ file:///C:/ci/coverage_1614614921360/work
cvxopt==1.2.7
cvxpy @ file:///D:/bld/cvxpy-split_1628845963166/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython @ file:///C:/ci/cython_1626257069493/work
dataclasses @ file:///tmp/build/80754af9/dataclasses_1614363715916/work
ecos @ file:///C:/ci/ecos_1614178312331/work
gast @ file:///tmp/build/80754af9/gast_1632115987108/work
grpcio @ file:///C:/ci/grpcio_1614884461394/work
html5lib==0.9999999
importlib-metadata @ file:///C:/ci/importlib-metadata_1631916885413/work
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
kiwisolver @ file:///C:/ci/kiwisolver_1612282446297/work
Markdown @ file:///C:/ci/markdown_1614364003005/work
matplotlib @ file:///D:/bld/matplotlib-base_1594091861366/work
nltk @ file:///tmp/build/80754af9/nltk_1632909634979/work
numpy @ file:///D:/bld/numpy_1626682185159/work
osqp @ file:///D:/bld/osqp_1628283302234/work
protobuf==3.18.1
publicsuffixlist==0.7.12
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
qdldl @ file:///D:/bld/qdldl-python_1628251902653/work
regex @ file:///C:/ci/regex_1629302333945/work
scipy @ file:///C:/bld/scipy_1629411592125/work
scs @ file:///D:/bld/scs_1628253679114/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
tornado @ file:///C:/ci/tornado_1606942379977/work
tqdm @ file:///tmp/build/80754af9/tqdm_1631818572807/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
Werkzeug==2.0.2
wincertstore==0.2
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
